+++ ft-sdk/CHANGELOG.md
@@ -66,6 +66,29 @@
 1. 支持开启 session replay 录制功能
 
 ---
+# agent 1.6.12
+1. 添加 `FTRUMConfig.setEnableTraceWebView` 配置是否开启通过 Android SDK 采集 WebView 数据,
+   通过 `FTRUMConfig.setAllowWebViewHost` 可以控制需要过滤 host 地址
+2. 添加 `ContentHandlerHelperEx.onExceptionWithFilter`, 可以针对本地网络错误进行过滤
+3. 添加 `FTSDKConfig.setRemoteConfiguration` 支持开启远程条件配置功能，
+   添加`FTSDKConfig.setRemoteConfigMiniUpdateInterval` 支持在开启远程控制之后，设置最短更新时间间隔
+
+---
+# agent 1.6.12-beta01
+1. 修正 webHost 空配置异常的问题
+
+---
+# agent 1.6.12-alpha02
+1. 添加 `FTSDKConfig.setRemoteConfiguration` 支持开启远程条件配置功能， 
+   添加`FTSDKConfig.setRemoteConfigMiniUpdateInterval` 支持在开启远程控制之后，设置最短更新时间间隔 
+
+---
+# agent 1.6.12-alpha01
+1. 添加 `FTRUMConfig.setEnableTraceWebView` 配置是否开启通过 Android SDK 采集 WebView 数据,
+   通过 `FTRUMConfig.setAllowWebViewHost` 可以控制需要过滤 host 地址
+2. 添加 `ContentHandlerHelperEx.onExceptionWithFilter`, 可以针对本地网络错误进行过滤
+
+---
 # agent 1.6.11
 1. 添加 `FTRUMConfig.setEnableTraceUserViewInFragment` 支持 fragment view 数据采集，默认为 false
 2. 添加 `FTSDKConfig.setLineDataModifier`、`FTSDKConfig.setDataModifier` 支持数据写入替换，支持数据脱敏